module-99-cui/pyproject.toml
[tool.poetry]
name = "module-0-cui"
version = "0.1.0"
description = ""
authors = ["abe.akira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "2.4.0"
aiohttp = "3.10.5"
aiosignal = "1.3.1"
aiosqlite = "0.20.0"
annotated-types = "0.7.0"
anyio = "4.4.0"
appnope = "0.1.4"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "2.4.1"
async-lru = "2.0.4"
attrs = "24.2.0"
babel = "2.16.0"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
comm = "0.2.2"
dataclasses-json = "0.6.7"
debugpy = "1.8.5"
decorator = "5.1.1"
defusedxml = "0.7.1"
distro = "1.9.0"
executing = "2.1.0"
fastjsonschema = "2.20.0"
fqdn = "1.5.1"
frozenlist = "1.4.1"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.2"
httpx-sse = "0.4.0"
idna = "3.8"
ipykernel = "6.29.5"
ipython = "8.27.0"
isoduration = "20.11.0"
jedi = "0.19.1"
jinja2 = "3.1.4"
jiter = "0.5.0"
json5 = "0.9.25"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
jupyter-events = "0.10.0"
jupyter-lsp = "2.2.5"
jupyter-client = "8.6.2"
jupyter-core = "5.7.2"
jupyter-server = "2.14.2"
jupyter-server-terminals = "0.5.3"
jupyterlab = "4.2.5"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.27.3"
langchain = "0.2.16"
langchain-community = "0.2.16"
langchain-core = "0.2.38"
langchain-openai = "0.1.23"
langchain-text-splitters = "0.2.4"
langgraph = "0.2.19"
langgraph-checkpoint = "1.0.9"
langgraph-checkpoint-sqlite = "1.0.3"
langgraph-sdk = "0.1.30"
langsmith = "0.1.117"
markupsafe = "2.1.5"
marshmallow = "3.22.0"
matplotlib-inline = "0.1.7"
mistune = "3.0.2"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
nbclient = "0.10.0"
nbconvert = "7.16.4"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
notebook = "7.2.2"
notebook-shim = "0.2.4"
numpy = "1.26.4"
openai = "1.44.1"
orjson = "3.10.7"
overrides = "7.7.0"
packaging = "24.1"
pandocfilters = "1.5.1"
parso = "0.8.4"
pexpect = "4.9.0"
platformdirs = "4.3.2"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.47"
psutil = "6.0.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pycparser = "2.22"
pydantic = "2.9.1"
pydantic-core = "2.23.3"
pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
python-json-logger = "2.0.7"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
referencing = "0.35.1"
regex = "2024.7.24"
requests = "2.32.3"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rpds-py = "0.20.0"
send2trash = "1.8.3"
setuptools = "74.1.2"
six = "1.16.0"
sniffio = "1.3.1"
soupsieve = "2.6"
sqlalchemy = "2.0.34"
stack-data = "0.6.3"
tavily-python = "0.4.0"
tenacity = "8.5.0"
terminado = "0.18.1"
tiktoken = "0.7.0"
tinycss2 = "1.3.0"
tornado = "6.4.1"
tqdm = "4.66.5"
traitlets = "5.14.3"
types-python-dateutil = "2.9.0.20240906"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
uri-template = "1.3.0"
urllib3 = "2.2.2"
wcwidth = "0.2.13"
webcolors = "24.8.0"
webencodings = "0.5.1"
websocket-client = "1.8.0"
wikipedia = "1.4.0"
yarl = "1.11.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
